requirements.txt
mavsdk==2.12.2
numpy==1.24.3
scipy==1.11.1
pyyaml==6.0
asyncio==3.4.3
pandas==2.0.3
matplotlib==3.7.2

